LUNA-25: Abstract Class for all of Luna's dialog windows
