+++ resources/lib/views/gameinfo.py
@@ -1,44 +1,18 @@
-import os
-
 import pyxbmct.addonwindow as pyxbmct
 
-from xbmcswift2 import xbmc, xbmcaddon, xbmcgui, Plugin
+from xbmcswift2 import xbmc, xbmcgui
 
-from resources.lib.di.requiredfeature import RequiredFeature
-
-_addon_path = xbmcaddon.Addon().getAddonInfo('path')
-
-COLOR_FO = '0xFFE0B074'
-COLOR_NF = '0xFF808080'
-COLOR_HEADING = '0xFFD6D6D6'
-COLOR_DETAILS = '0xFF707070'
-COLOR_SELECTED = '0xFFF1F1F1'
+from resources.lib.views.abstractdialogwindow import AbstractDialogWindow
 
 
-class GameInfo(pyxbmct.AddonDialogWindow):
-    plugin = RequiredFeature('plugin')
-    core = RequiredFeature('core')
-
+class GameInfo(AbstractDialogWindow):
     def __init__(self, game, title=''):
         super(GameInfo, self).__init__(title)
         self.game = game
-        background = None
-        if self.core.get_active_skin() == 'skin.osmc':
-            media_path = '/usr/share/kodi/addons/skin.osmc/media'
-            if os.path.exists(media_path):
-                background = os.path.join(media_path, 'dialogs/DialogBackground_old.png')
 
-        if background is not None:
-            self.background.setImage(background)
-            self.removeControl(self.title_background)
-            self.removeControl(self.window_close_button)
-            self.removeControl(self.title_bar)
-
-        self.setGeometry(1280, 720, 12, 6, padding=60)
         self.set_info_controls(game)
-        self.set_active_controls(game)
+        self.set_active_controls()
         self.set_navigation()
-        self.connect(pyxbmct.ACTION_NAV_BACK, self.close)
         # init controls
         self.image = None
         self.genre = None
@@ -49,7 +23,8 @@ class GameInfo(pyxbmct.AddonDialogWindow):
         self.button_fanart = None
 
     def set_info_controls(self, game):
-        title_label = pyxbmct.Label(game.name, alignment=pyxbmct.ALIGN_LEFT, font='XLarge', textColor=COLOR_HEADING)
+        title_label = pyxbmct.Label(game.name, alignment=pyxbmct.ALIGN_LEFT, font='XLarge',
+                                    textColor=self.COLOR_HEADING)
         self.placeControl(title_label, 0, 0, 2, 3)
 
         image_path = game.get_selected_poster()
@@ -58,13 +33,13 @@ class GameInfo(pyxbmct.AddonDialogWindow):
         self.image = pyxbmct.Image(image_path)
         self.placeControl(self.image, 2, 0, 6, 1)
 
-        genre_label = pyxbmct.Label('Genre', alignment=pyxbmct.ALIGN_LEFT, font='Med', textColor=COLOR_DETAILS)
+        genre_label = pyxbmct.Label('Genre', alignment=pyxbmct.ALIGN_LEFT, font='Med', textColor=self.COLOR_DETAILS)
         self.placeControl(genre_label, 2, 2)
         self.genre = pyxbmct.FadeLabel(_alignment=pyxbmct.ALIGN_LEFT, font='Med')
         self.placeControl(self.genre, 2, 3, columnspan=3)
         self.genre.addLabel(game.get_genre_as_string())
 
-        year_label = pyxbmct.Label('Year', alignment=pyxbmct.ALIGN_LEFT, font='Med', textColor=COLOR_DETAILS)
+        year_label = pyxbmct.Label('Year', alignment=pyxbmct.ALIGN_LEFT, font='Med', textColor=self.COLOR_DETAILS)
         self.placeControl(year_label, 3, 2)
         self.year = pyxbmct.Label(game.year, alignment=pyxbmct.ALIGN_LEFT, font='Med')
         self.placeControl(self.year, 3, 3)
@@ -74,20 +49,20 @@ class GameInfo(pyxbmct.AddonDialogWindow):
         self.plot.setText(game.plot)
         self.plot.autoScroll(delay=5000, time=2000, repeat=10000)
 
-    def set_active_controls(self, game):
-        self.button_play = pyxbmct.Button('Play', focusTexture='', noFocusTexture='', focusedColor=COLOR_FO,
-                                          textColor=COLOR_NF, font='Med', alignment=pyxbmct.ALIGN_LEFT)
+    def set_active_controls(self):
+        self.button_play = pyxbmct.Button('Play', focusTexture='', noFocusTexture='', focusedColor=self.COLOR_FO,
+                                          textColor=self.COLOR_NF, font='Med', alignment=pyxbmct.ALIGN_LEFT)
         self.placeControl(self.button_play, 11, 0)
         self.connect(self.button_play, self.launch_game)
 
         self.button_cover_art = pyxbmct.Button('Choose Cover Art', focusTexture='', noFocusTexture='',
-                                               focusedColor=COLOR_FO, textColor=COLOR_NF, font='Med',
+                                               focusedColor=self.COLOR_FO, textColor=self.COLOR_NF, font='Med',
                                                alignment=pyxbmct.ALIGN_LEFT)
         self.placeControl(self.button_cover_art, 11, 1, columnspan=2)
         self.connect(self.button_cover_art, self.select_cover_art)
 
         self.button_fanart = pyxbmct.Button('Choose Fanart', focusTexture='', noFocusTexture='',
-                                            focusedColor=COLOR_FO, textColor=COLOR_NF, font='Med',
+                                            focusedColor=self.COLOR_FO, textColor=self.COLOR_NF, font='Med',
                                             alignment=pyxbmct.ALIGN_LEFT)
         self.placeControl(self.button_fanart, 11, 3, columnspan=2)
         self.connect(self.button_fanart, self.select_fanart)
@@ -122,11 +97,3 @@ class GameInfo(pyxbmct.AddonDialogWindow):
             self.core.get_storage().sync()
             self.image = pyxbmct.Image(browser)
             self.placeControl(self.image, 2, 0, 6, 1)
-
-    def setAnimation(self, control):
-        control.setAnimations(
-                [
-                    ('WindowOpen', 'effect=fade start=0 end=100 time=500',),
-                    ('WindowClose', 'effect=fade start=100 end=0 time=500',)
-                ]
-        )
